+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - ?
 
 - update config file to amnesia.xml
+- add more admin options
 
 ## [0.3.0] - 2022-07-02